+++ src/lsp/cobol_typeck/cobol_typeck.ml
@@ -23,3 +23,6 @@ module Outputs = Typeck_outputs
 module Diagnostics = Typeck_diagnostics
 module Results = Typeck_results
 include Typeck_engine
+
+(* Additional: *)
+module References = Typeck_references