+++ test/testing_utils/prog_typeck.ml
@@ -12,7 +12,7 @@
 (**************************************************************************)
 
 let typeck ?parser_options ?source_format ?filename contents =
-  let fold_exec_block' ~register_name:_ _exec_block acc =
+  let fold_exec_block' ~data_definitions:_ _exec_block acc =
     (* We could use Superbol_preprocs.Esql.fold_exec_block' to test
        SQL statements*)
     acc